Extract PagerAdapter
